services/inference/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "observe"
version = "0.1.0"
description = "Model inference service for Coach Kata"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "coachkata-libraries",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "mediapipe>=0.10.0",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.62.3",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/observe"] 